password change / reset functionality, admin restrictions added

+++ src/main/java/it/bz/beacon/api/service/user/UserService.java
@@ -2,12 +2,12 @@ package it.bz.beacon.api.service.user;
 
 import it.bz.beacon.api.db.model.User;
 import it.bz.beacon.api.db.repository.UserRepository;
-import it.bz.beacon.api.exception.db.DuplicateEntryException;
+import it.bz.beacon.api.exception.auth.InsufficientRightsException;
+import it.bz.beacon.api.exception.auth.InvalidPasswordException;
 import it.bz.beacon.api.exception.db.AuthenticatedUserNotDeletableException;
+import it.bz.beacon.api.exception.db.DuplicateEntryException;
 import it.bz.beacon.api.exception.db.UserNotFoundException;
-import it.bz.beacon.api.model.BaseMessage;
-import it.bz.beacon.api.model.UserCreation;
-import it.bz.beacon.api.model.UserUpdate;
+import it.bz.beacon.api.model.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -36,21 +36,28 @@ public class UserService implements IUserService {
     }
 
     @Override
-    public User create(UserCreation userCreation) {
-        try {
-            userCreation.setPassword(passwordEncoder.encode(userCreation.getPassword()));
-            return repository.save(User.create(userCreation));
-        } catch (DataIntegrityViolationException e) {
-            throw new DuplicateEntryException();
+    public User create(UserCreation userCreation) throws DuplicateEntryException, InsufficientRightsException {
+        if (isAdmin()) {
+            try {
+                userCreation.setPassword(passwordEncoder.encode(userCreation.getPassword()));
+                return repository.save(User.create(userCreation));
+            } catch (DataIntegrityViolationException e) {
+                throw new DuplicateEntryException();
+            }
         }
+
+        throw new InsufficientRightsException();
     }
 
     @Override
     public User update(long id, UserUpdate userUpdate) throws UserNotFoundException {
+        User authUser = (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
         return repository.findById(id).map(user -> {
-            if (userUpdate.getPassword() != null) {
-                userUpdate.setPassword(passwordEncoder.encode(userUpdate.getPassword()));
+            if (!isAdmin() || !authUser.getUsername().equals(user.getUsername())) {
+                throw new InsufficientRightsException();
             }
+
             user.applyUpdate(userUpdate);
 
             return repository.save(user);
@@ -58,18 +65,65 @@ public class UserService implements IUserService {
     }
 
     @Override
-    public BaseMessage delete(long id) throws UserNotFoundException, AuthenticatedUserNotDeletableException {
-        User authUser = (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+    public BaseMessage delete(long id) throws UserNotFoundException, AuthenticatedUserNotDeletableException,
+            InsufficientRightsException {
 
+        if (isAdmin()) {
+            User authUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+            return repository.findById(id).map(
+                    user -> {
+                        if (user.getId().longValue() == authUser.getId().longValue()) {
+                            throw new AuthenticatedUserNotDeletableException();
+                        }
+                        repository.delete(user);
+
+                        return new BaseMessage("User deleted");
+                    }
+            ).orElseThrow(UserNotFoundException::new);
+        }
+
+        throw new InsufficientRightsException();
+    }
+
+    @Override
+    public BaseMessage resetPassword(long id, PasswordReset passwordReset) throws UserNotFoundException,
+            InsufficientRightsException {
+        if (isAdmin()) {
+            return repository.findById(id).map(
+                    user -> {
+                        user.setPassword(passwordEncoder.encode(passwordReset.getNewPassword()));
+                        repository.save(user);
+
+                        return new BaseMessage("Password reset");
+                    }
+            ).orElseThrow(UserNotFoundException::new);
+        }
+
+        throw new InsufficientRightsException();
+    }
+
+    @Override
+    public BaseMessage changePassword(long id, PasswordChange passwordChange) throws InvalidPasswordException,
+            UserNotFoundException {
         return repository.findById(id).map(
                 user -> {
-                    if (user.getId().longValue() == authUser.getId().longValue()) {
-                        throw new AuthenticatedUserNotDeletableException();
+                    if (!passwordEncoder.matches(passwordChange.getOldPassword(), user.getPassword())) {
+                        throw new InvalidPasswordException();
                     }
-                    repository.delete(user);
 
-                    return new BaseMessage("User deleted");
+                    user.setPassword(passwordEncoder.encode(passwordChange.getNewPassword()));
+                    repository.save(user);
+                    return new BaseMessage("Password changed");
                 }
         ).orElseThrow(UserNotFoundException::new);
+    }
+
+    private boolean isAdmin() {
+        try {
+            return ((User)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).isAdmin();
+        } catch (Exception e) {
+            return false;
+        }
     }
 }
